Moves all code to /src folder
