+++ src/cameraControl.h
