+++ src/detectJoystick.cpp
