+++ src/stdSocketTools.cpp
